demo/src/opentelemetry/requirements.txt
flask
requests
pymysql == 0.9.3
opentelemetry-sdk == 0.10b0
opentelemetry-ext-flask == 0.10b0
opentelemetry-ext-requests == 0.10b0
opentelemetry-ext-prometheus == 0.10b0
opentelemetry-ext-wsgi == 0.10b0
opentelemetry-ext-pymysql == 0.10b0
opentelemetry-azure-monitor == 0.4b0